server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.3
pytest==7.4.0
pytest-flask==1.2.0